pyproject.toml
[tool.poetry]
name = "minly-back-python"
version = "0.1.0"
description = ""
authors = ["Evgeniy Smirnov <59017716+mrlimax1@users.noreply.github.com>"]
packages = [
    { include = "src" },
    { include = "src/**/*.py" },
]
[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.78.0"
pydantic = "^1.9.1"
pony = "^0.7.16"
gunicorn = "^20.1.0"
uvicorn = "^0.17.6"
pip = "^22.1.2"
python-decouple = "^3.6"
psycopg2-binary = "^2.9.3"
aiohttp = "^3.8.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "src.main:start"
